Remove node_modules from repo
